feat(env): configure webpack, add glsl (raw) loader
