+++ src/index.html
@@ -5,13 +5,13 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="stylesheet" href="style.css">
-  <title>Document</title>
+  <title>WebGL test</title>
 </head>
-<body onload="start()">
-  <script src="https://mdn.github.io/webgl-examples/tutorial/gl-matrix.js"></script>
-  <script src="main.js"></script>
+<body>
   <div id="app">
     <canvas id="gl-canvas" width=1024 height=560></canvas>
   </div>
+  <script src="https://mdn.github.io/webgl-examples/tutorial/gl-matrix.js"></script>
+  <script src="bundle.js"></script>
 </body>
 </html>
