+++ src/shaders/fragment.glsl
@@ -1,0 +1,3 @@
+void main() {
+  gl_FragColor = vec4(1.0, 1.0, 1.0, 1.0);
+}
